Implement wallet backups for redb
